LM2HM bugfix, Thanks Cecilia

+++ martin_LMtoHM.m
@@ -97,7 +97,9 @@ end
 
 if hm_overlays
     im=imread([pathname,namebase,'_im.tif']);
-    om=imread([pathname,namebase,'_om.tif']);
+    if exist([pathname,namebase,'_om.tif'],'file')
+        om=imread([pathname,namebase,'_om.tif']);
+    end
 end
 smf1=smf;
 lm=imadjust(lm);
@@ -242,9 +244,12 @@ imwrite(tfmcircle,[outfileroot,file,'_hm_prediction.tif'],'Compression','none');
 
 if hm_overlays
  [im2 xdata ydata]=imtransform(gm,thm,'FillValues',128,'XData', [1 size(hm,2)],'YData',[1 size(hm,1)],'Size',size(hm));
- [om2 xdata ydata]=imtransform(rm,thm,'FillValues',128,'XData', [1 size(hm,2)],'YData',[1 size(hm,1)],'Size',size(hm));  
+ if exist([pathname,namebase,'_om.tif'],'file')
+     [om2 xdata ydata]=imtransform(rm,thm,'FillValues',128,'XData', [1 size(hm,2)],'YData',[1 size(hm,1)],'Size',size(hm));  
+     imwrite(om2,[outfileroot,file,'_hm_om.tif'],'Compression','none');
+ end
  imwrite(im2,[outfileroot,file,'_hm_im.tif'],'Compression','none');
- imwrite(om2,[outfileroot,file,'_hm_om.tif'],'Compression','none');
+
 end
 
 impred=tfmcircle+uint8(sm/255);
